th.sh: Introduce gauges with baseline and other changes
Change the way the script decides when to update baseline. Previously it
was either ever tick or never. When never was selected, the units were
always bytes.

Instead decouple the two: make it possible to have gauges that measure
something else than bytes. B:, P: and uS: still select units. G: was
dropped, and instead N: was introduce to choose to update never. B: N: is
exactly what G: used to be.

Additionally, introduce a toggle T:, which selects update each tick (which
was previously implied in all modes except G:); and O:, which updates
baseline once, and then never again (so that one gets counter value
relative to the time that the tool was started). Also introduce 1: to just
show unitless value.

Change B: to always mean bytes, introduce Bb: to mean bytes that should be
shown as bits.

Signed-off-by: Petr Machata <petrm@mellanox.com>